refactor(application): move skill modules into submodule
