+++ src-tauri/src/application/mod.rs
@@ -2,14 +2,7 @@ mod adapter_service;
 #[cfg(test)]
 mod critical_paths_suite;
 mod mcp;
-mod skill_github_repository;
-mod skill_listing_service;
-mod skill_metadata_parser;
-mod skill_mutation_path_resolver;
-mod skill_mutation_payload;
-mod skill_mutation_service;
-mod skill_path_resolver;
-mod skill_repository_discovery_service;
+mod skill;
 
 pub use adapter_service::AdapterService;
-pub use skill_repository_discovery_service::SkillRepositoryDiscoveryService;
+pub use skill::repository_discovery_service::SkillRepositoryDiscoveryService;
