+++ src-tauri/src/application/skill/metadata_parser.rs
