+++ src-tauri/src/application/skill/mod.rs
@@ -1,0 +1,8 @@
+pub(super) mod github_repository;
+pub(super) mod listing_service;
+pub(super) mod metadata_parser;
+pub(super) mod mutation_path_resolver;
+pub(super) mod mutation_payload;
+pub(super) mod mutation_service;
+pub(super) mod path_resolver;
+pub(super) mod repository_discovery_service;
