+++ src-tauri/src/application/skill/path_resolver.rs
