+++ src-tauri/src/application/skill/repository_discovery_service.rs
@@ -3,7 +3,7 @@ use crate::contracts::{
     skill_discovery::{DiscoverSkillRepositoryResponse, DiscoveredSkillCandidate},
 };
 
-use super::skill_github_repository::scan_github_repository;
+use super::github_repository::scan_github_repository;
 
 pub struct SkillRepositoryDiscoveryService;
